==> analysis/01_exploreArrayBiasEPIC.Rmd <==
---
title: 'Gene set testing for Illumina HumanMethylation Arrays'
subtitle: "Exploring the biases on EPIC arrays "
author: "Jovana Maksimovic, Alicia Oshlack and Belinda Phipson"
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
options(scipen=10000)
```

```{r, message=FALSE, warning=FALSE, results='hide'}
library(here)
library(glue)
library(minfi)
library(IlluminaHumanMethylationEPICanno.ilm10b4.hg19)
library(IlluminaHumanMethylation450kanno.ilmn12.hg19)
library(missMethyl)
library(org.Hs.eg.db)
library(GO.db)
library(patchwork)
library(grid)
library(ggplot2)
library(tibble)
library(dplyr)
source(here("code/utility.R"))
```

## Array properties

Get the array annotation data.

```{r}
ann <- loadAnnotation(arrayType="EPIC")
```

Associate CpGs to genes (ENTREZ ID) using the Illumina annotation information.

```{r}
flatAnn <- loadFlatAnnotation(ann)
```

The number of CpGs annotated to a gene is highly variable.

```{r, fig.height=10, fig.width=8}
numCpgsPerGene <- as.vector(table(flatAnn$entrezid))
summary(numCpgsPerGene)
```

```{r}
dat <- data.frame(table(table(flatAnn$entrezid)))
numCpgsPerGene <- as.vector(table(flatAnn$entrezid))
med <- median(numCpgsPerGene)
mod <- getMode(numCpgsPerGene)

p <- ggplot(dat, aes(x=Var1, y=Freq)) +
    geom_segment(aes(x=Var1, xend=Var1, y=0, yend=Freq), color="darkgrey") +
    geom_point( color="black", size=1) +
    geom_vline(xintercept = med, linetype = "dashed", color = "red") +
    annotate("text", x = med + 2, label=glue("median = {med}"), y=720, colour="red", 
             size = 3, hjust="left") +
    geom_vline(xintercept = mod, linetype = "dashed", color = "blue") +
    annotate("text", x = mod + 2, label=glue("mode = {mod}"), y=780, colour="blue", 
             size = 3, hjust="left") +
    scale_x_discrete(breaks = c(1,10,20,30,45,55,70,80,90,100,120,135,150,160,180,
                                200,215,250,330,429,1485)) +
    theme(axis.text.x=element_text(angle=45, hjust=1)) +
    xlab("No. CpGs per gene") +
    ylab("Frequency")

p
```

Save figure for use in manuscript.

```{r}
outDir <- here::here("output/figures")
if (!dir.exists(outDir)) dir.create(outDir)

fig <- here("output/figures/Fig-1A.rds")
saveRDS(p, fig, compress = FALSE)
```

The number of genes a CpG maps to can also vary, although the majority of CpGs only map to one gene.

```{r}
dat <- data.frame(table(table(flatAnn$cpg)))
dat$Split <- ifelse(dat$Freq > 2500, "A", 
                    ifelse(dat$Freq < 2500 & dat$Freq > 65,"B","C"))

a <- ggplot(dat[dat$Split == "A",], aes(x=Var1, y=Freq)) +
    geom_bar(stat = "identity") +
    theme(axis.title.x = element_blank(), 
          axis.text.y=element_text(angle=90, hjust=0.5)) +
    ylab("Frequency")

b <- ggplot(dat[dat$Split == "B",], aes(x=Var1, y=Freq)) +
    geom_bar(stat = "identity") +
    theme(axis.title.x=element_blank(),
          axis.title.y = element_blank())

c <- ggplot(dat[dat$Split == "C",], aes(x=Var1, y=Freq)) +
    geom_bar(stat = "identity") +
    theme(axis.title.y = element_blank(),
          axis.title.x = element_blank()) 

p <- a + b + c + plot_layout(widths = c(1,2,6)) + 
   plot_annotation(caption = "No. genes mapped to a CpG",
                   theme = theme(plot.caption = element_text(hjust = 0.5,
                                                             size = 12)))
p
```

Save figure for use in manuscript.

```{r}
fig <- here("output/figures/Fig-1D.rds")
saveRDS(p, fig, compress = FALSE)
```

Explore the distribution of the average number of CpGs per gene, per GO category. 

```{r}
ann450 <- loadAnnotation("450k")
flatAnn450 <- loadFlatAnnotation(ann450)

cpgEgGoEPIC <- cpgsEgGoFreqs(flatAnn)
cpgEgGoEPIC$Array <- "EPIC"

cpgEgGo450 <- cpgsEgGoFreqs(flatAnn450)
cpgEgGo450$Array <- "450k"

cpgEgGoEPIC %>% bind_rows(cpgEgGo450) %>% 
    group_by(Array, GO) %>%
    summarise(avg = mean(Freq)) -> datSum

maxEPIC <- max(datSum$avg[datSum$Array == "EPIC"])
max450 <- max(datSum$avg[datSum$Array == "450k"])

bw <- 2
pal <- paletteer::paletteer_d("RColorBrewer::Set1", 2)
platform <- c("EPIC" = pal[1], "450k" = pal[2])
p <- ggplot(datSum, aes(x=avg)) +
    geom_histogram(dat = subset(datSum, Array == "450k"),
                   alpha = 0.5, aes(fill = "450k"), binwidth = bw) +
    geom_histogram(dat = subset(datSum, Array == "EPIC"),
                   alpha = 0.5, aes(fill = "EPIC"), binwidth = bw) +
    geom_vline(xintercept = maxEPIC, linetype="dashed", colour = pal[1]) +
    geom_vline(xintercept = max450, linetype="dashed", colour = pal[2]) +
    labs(x="Mean no. CpGs per gene per GO category", y = "Frequency",
         fill = "Platform") +
    annotate("text", x = maxEPIC - 4, label=glue("Max. EPIC = {maxEPIC}"),
             y = 1500, colour=pal[1], size = 2.5, hjust="right") +
    annotate("text", x = max450 - 4, label=glue("Max. 450k = {max450}"),
             y = 2500, colour=pal[2], size = 2.5, hjust="right") +
    scale_fill_manual(values = platform)
p
```

Save figure for use in manuscript.

```{r}
fig <- here("output/figures/Fig-1C.rds")
saveRDS(p, fig, compress = FALSE)
```

## Null simulations

We randomly select 50, 100, 500, 1000, 5000 and 10000 sets of CpGs and perform GO testing on each set 100 times, with and without adjusting for the various biases on the array. 

The code used to produce the simulation results can be found in the `code/random-cpg-sims` directory. It consists of three scripts: `genRandCpgSimJobs.R`, `randomCpgSim.R` and `processRandCpgSim.R`. The `genRandCpgSimJobs.R` script creates and submits [Slurm](https://slurm.schedmd.com/overview.html) job scripts that run the `randomCpgSim.R` script, in parallel, on a HPC. Each job executes *one* of the 100 simulations, for a fixed number of randomly selected CpGs, using either the 450K or EPIC array annotation. The results of each job are saved as an RDS file named `{arrayType}.{noCpgs}.{simNo}.rds` in the `output/random-cpg-sims` directory. Once all simulation jobs are complete, the `processRandCpgSim.R` must be executed to collate the results into a single object for each array type, which are then saved as `450K.rds` and `EPIC.rds` in the `output/random-cpg-sims` directory. The intermediate RDS files are moved into `output/random-cpg-sims/.bin`, which can then be deleted, if no longer required. The subsequent section requires `EPIC.rds` to be present in the `output/random-cpg-sims` directory for downstream analysis and plotting.             

### Results

The following boxplots show what proportion of the 100 simulations, at each level of CpGs sampled, had a raw p-value less than 0.05. This gives us an idea of the false discovery rate with and without adjustment for the number of CpGs annotated to a gene.

```{r}
dat <- readRDS(here("output/random-cpg-sims/EPIC.rds"))

dat %>% filter(method %in% names(dict)) %>%
    mutate(method = unname(dict[method])) %>%
    group_by(simNo, noCpgs, method) %>% 
    summarise(pSig = sum(P.DE < 0.05)/length(P.DE)) -> sigDat

p <- ggplot(sigDat, aes(x=noCpgs, y=pSig, fill=method)) + 
    geom_violin(size = 0.3) +
    geom_hline(yintercept=0.05, linetype="dashed", color = "red") +
    labs(y="Prop. GO cat. with p-value < 0.05", x="No. randomly sampled CpGs",
         fill="Method") +
    scale_fill_manual(values = methodCols)  +
    facet_grid(cols = vars(noCpgs), scales = "free_x") + 
    theme(strip.background = element_blank(), 
          strip.text = element_blank())
p
```

```{r}
fig <- here("output/figures/Fig-3A.rds")
saveRDS(p, fig, compress = FALSE)
```

QQ plots of randomly selected simulations at each level of CpGs sampled.

```{r, fig.height=10, fig.width=7}
set.seed(42)
s <- sample(1:100, 3)
dat %>% filter(simNo %in% s) %>%
    filter(method %in% names(dict)) %>%
    mutate(method = unname(dict[method])) %>%
    arrange(simNo, noCpgs, method, P.DE) %>%
    group_by(simNo, noCpgs, method) %>%
    mutate(exp = 1:n()/n()) -> subDat

p <- ggplot(subDat, aes(x=-log10(exp), y=-log10(P.DE), color=method)) + 
    geom_point(shape = 1, size = 0.5) + 
    facet_grid(noCpgs ~ simNo, scales = "free_y") +
    scale_color_manual(values = methodCols)

p + geom_line(aes(x=-log10(exp), y=-log10(exp)),
              linetype="dashed", color = "black") +
    labs(y=expression(Observed~~-log[10](italic(p))), 
         x=expression(Expected~~-log[10](italic(p))),
         color="Method") +
    theme(legend.position="bottom", strip.text.x = element_blank())
```

Explore the relationship between the median, average number of CpGs, per gene, per GO category and the various sources of bias on the array.

```{r}
goFreq <- as_tibble(unique(cpgEgGoEPIC[,c("GO","Freq")]))

dat %>% filter(method %in% names(dict)) %>%
    mutate(method = unname(dict[method])) %>% 
    filter(P.DE < 0.05) %>%
    inner_join(goFreq, by=c("GO" = "GO")) %>%
    group_by(simNo, noCpgs, method, GO) %>%
    summarise(avgFreq=mean(Freq)) %>% 
    group_by(simNo, noCpgs, method) %>%
    summarise(medAvgFreq=median(avgFreq)) -> medAvgDat

goFreq %>% group_by(GO) %>% 
    summarise(med = median(Freq)) %>%
    dplyr::select(med) %>%
    summarise(med = median(med)) -> medPerGO

dat_text <- data.frame(
  label = c(rep("",4),
            "Med. Med. # Genes", 
            glue("per GO Cat. = {round(medPerGO$med, 2)}")),
  method = "GOmeth",
  noCpgs = unique(medAvgDat$noCpgs),
  x = unique(medAvgDat$noCpgs),
  y = rep(medPerGO$med + 1, 6)
)

p <- ggplot(medAvgDat, aes(x=noCpgs, y=medAvgFreq, fill=method)) +
    geom_violin(size = 0.3, scale = "area") +
    stat_summary(geom="point", size = 0.5, color="white",
                 position = position_dodge(0.9),
                 show.legend = FALSE, fun = median) +
    labs(y="Median avg. no. CpGs/gene/GO cat.",
         x="No. randomly sampled CpGs",
         fill="Method")  +
    scale_fill_manual(values = methodCols) +
    geom_hline(yintercept = medPerGO$med, linetype="dashed", color = "red") +
    geom_text(data = dat_text, colour="red", size = 2.5, hjust="left", 
              mapping = aes(x = -Inf, y = y, label = label)) +
    facet_grid(cols = vars(noCpgs), scales = "free_x") +
    theme(strip.background = element_blank(),
          strip.text = element_blank())
p
```

Save figure for use in manuscript.

```{r}
fig <- here("output/figures/Fig-3B.rds")
saveRDS(p, fig, compress = FALSE)
```
