==> analysis/06_runTimeComparison.Rmd <==
---
title: 'Gene set testing for Illumina HumanMethylation Arrays'
subtitle: "Comparing the run-time of different methods"
author: "Jovana Maksimovic, Alicia Oshlack and Belinda Phipson"
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, message=FALSE, warning=FALSE, results='hide'}
library(here)
library(minfi)
library(limma)
library(reshape2)
library(missMethyl)
library(ggplot2)
library(glue)
library(tidyverse)
library(patchwork)
library(ChAMPdata)
library(tictoc)
library(IlluminaHumanMethylationEPICanno.ilm10b4.hg19)
library(EnsDb.Hsapiens.v75)
library(ChAMP)
library(ebGSEA)
library(methylGSA)
library(gt)
source(here("code/utility.R"))
```

# Compare run-time of different methods

## Setup inputs

Create database for translating gene IDs.

```{r}
edb <- EnsDb.Hsapiens.v75
transIDs <- genes(edb, columns = c("symbol", "gene_id", "entrezid"), 
                       return.type = "DataFrame")
```

## Measure run-times for different methods

Execute and record run-time for each method (on a single core) for the three different contrasts.

```{r, message=FALSE}
outDir <- here::here("data/cache-runtime")
if (!dir.exists(outDir)) dir.create(outDir)

inFile <- here("data/cache-runtime/run-time-results.rds")

if(!file.exists(inFile)){
    
    data("PathwayList")
    keep <- sapply(PathwayList, function(x) any(x %in% transIDs$symbol))
    symbol <- suppressMessages(lapply(PathwayList[keep], function(x){
        tmp <- unlist(transIDs$symbol[transIDs$symbol %in% x], use.names = FALSE)
        tmp[!is.na(tmp)]
    }))
    entrezid <- suppressMessages(lapply(symbol, function(x){
        tmp <- unlist(transIDs$entrezid[transIDs$symbol %in% x], use.names = FALSE)
        tmp[!is.na(tmp)]
    }))
    
    load(here("data/cache-intermediates/input.RData"))
    bVals <- ilogit2(mVals)
    anno <- minfi::getAnnotation(IlluminaHumanMethylationEPICanno.ilm10b4.hg19)
    timing <- NULL
    
    for(i in 1:ncol(tfit$contrasts)){
        top <- topTreat(tfit, coef = i, number = 5000)
        
        tic("gometh")
        res <- gsameth(sig.cpg = rownames(top), 
                       all.cpg = rownames(tfit$coefficients), collection = entrezid, 
                       array.type = "EPIC", anno = anno)
        toc(log = TRUE, quiet = TRUE)
        tmp <- strsplit2(tic.log(format = TRUE)[[1]], " ")
        log <- data.frame(method = gsub(":", "", tmp[1]), time = tmp[2], 
                          contrast = colnames(tfit$contrasts)[i])
        tic.clearlog()
        timing <- bind_rows(timing, log)
        
        tic("mgsa.glm")
        res <- methylglm(cpg.pval = tfit$p.value[,i],
                         FullAnnot = anno, minsize = minsize, maxsize = maxsize,
                         GS.list = symbol, GS.idtype = "SYMBOL")
        toc(log = TRUE, quiet = TRUE)
        tmp <- strsplit2(tic.log(format = TRUE)[[1]], " ")
        log <- data.frame(method = gsub(":", "", tmp[1]), time = tmp[2], 
                          contrast = colnames(tfit$contrasts)[i])
        tic.clearlog()
        timing <- bind_rows(timing, log)
        
        tic("mgsa.ora")
        res <- methylRRA(cpg.pval = tfit$p.value[,i],
                         method = "ORA", FullAnnot = anno, minsize = minsize,
                         maxsize = maxsize, GS.list = symbol,
                         GS.idtype = "SYMBOL")
        toc(log = TRUE, quiet = TRUE)
        tmp <- strsplit2(tic.log(format = TRUE)[[1]], " ")
        log <- data.frame(method = gsub(":", "", tmp[1]), time = tmp[2], 
                          contrast = colnames(tfit$contrasts)[i])
        tic.clearlog()
        timing <- bind_rows(timing, log)
        
        tic("mgsa.gsea")
        res <- methylRRA(cpg.pval = tfit$p.value[,i],
                         method = "GSEA", FullAnnot = anno, minsize = minsize,
                         maxsize = maxsize, GS.list = symbol,
                         GS.idtype = "SYMBOL")
        toc(log = TRUE, quiet = TRUE)
        tmp <- strsplit2(tic.log(format = TRUE)[[1]], " ")
        log <- data.frame(method = gsub(":", "", tmp[1]), time = tmp[2], 
                          contrast = colnames(tfit$contrasts)[i])
        tic.clearlog()
        timing <- bind_rows(timing, log)
        
        cellType <- names(tfit$contrasts[,i])[tfit$contrasts[,i] != 0]
        tic("ebgsea")
        #tic("champ.ebgsea")
        # ebgs <- champ.ebGSEA(beta = mVals[,targets$CellType %in% cellType],
        #                      pheno = targets$CellType[targets$CellType %in% cellType],
        #                      minN = 5, adjPval=1, arraytype = "EPIC")
        groups <- names(tfit$contrasts[,i])[tfit$contrasts[,i] != 0]
        samps <- as.logical(unname(rowSums(tfit$design[,groups])))
        pheno <- unname(tfit$design[samps, groups][,1])
        
        gtRanks <- doGT(pheno.v = pheno,
                        data.m = bVals[, samps],
                        array = "EPIC",
                        ncores = 1)
        ebgs <- data.frame(doGSEAwt(rankEID.m = gtRanks, ptw.ls = entrezid,
                                    ncores = 1, minN = minsize, adjPVth = 1)$`Rank(P)`)
        
        toc(log = TRUE, quiet = TRUE)
        tmp <- strsplit2(tic.log(format = TRUE)[[1]], " ")
        log <- data.frame(method = gsub(":", "", tmp[1]), time = tmp[2], 
                          contrast = colnames(tfit$contrasts)[i])
        tic.clearlog()
        timing <- bind_rows(timing, log)
        
    }
    saveRDS(timing, file = inFile)
    
} else {
    timing <- readRDS(inFile)
    colnames(timing)[3] <- "contrast"
    
}
```

## Visualise results

Plot run-time results.

```{r}
timing %>% mutate(time = as.integer(time)) %>%
    mutate(method = unname(dict[method])) -> dat

p1 <- ggplot(dat, aes(x = reorder(method, -time), y = time/60, 
                      fill = contrast)) + 
    geom_bar(position = "dodge", stat = "identity") +
    labs(x = "Method", y = "Run time (minutes)", fill = "Contrast")
p1
```

```{r}
inFile <- here("data/cache-runtime/run-time-mcores-ebgsea.rds")

if(!file.exists(inFile)){
    
    data("PathwayList")
    keep <- sapply(PathwayList, function(x) any(x %in% transIDs$symbol))
    symbol <- suppressMessages(lapply(PathwayList[keep], function(x){
        tmp <- unlist(transIDs$symbol[transIDs$symbol %in% x], use.names = FALSE)
        tmp[!is.na(tmp)]
    }))
    entrezid <- suppressMessages(lapply(symbol, function(x){
        tmp <- unlist(transIDs$entrezid[transIDs$symbol %in% x], use.names = FALSE)
        tmp[!is.na(tmp)]
    }))
    
    load(here("data/cache-intermediates/input.RData"))
    bVals <- ilogit2(mVals)
    anno <- minfi::getAnnotation(IlluminaHumanMethylationEPICanno.ilm10b4.hg19)
    multi <- NULL
    
    for(i in 1:ncol(tfit$contrasts)){
        
        tic("ebgsea.10")
        groups <- names(tfit$contrasts[,i])[tfit$contrasts[,i] != 0]
        samps <- as.logical(unname(rowSums(tfit$design[,groups])))
        pheno <- unname(tfit$design[samps, groups][,1])
        
        gtRanks <- doGT(pheno.v = pheno,
                        data.m = bVals[, samps],
                        array = "EPIC",
                        ncores = 10)
        ebgs <- data.frame(doGSEAwt(rankEID.m = gtRanks, ptw.ls = entrezid,
                                    ncores = 10, minN = minsize, adjPVth = 1)$`Rank(P)`)
        toc(log = TRUE, quiet = TRUE)
        tmp <- strsplit2(tic.log(format = TRUE)[[1]], " ")
        log <- data.frame(method = gsub(":", "", tmp[1]), time = tmp[2], 
                          contrast = colnames(tfit$contrasts)[i])
        tic.clearlog()
        multi <- bind_rows(multi, log)
        
        tic("ebgsea.20")
        groups <- names(tfit$contrasts[,i])[tfit$contrasts[,i] != 0]
        samps <- as.logical(unname(rowSums(tfit$design[,groups])))
        pheno <- unname(tfit$design[samps, groups][,1])
        
        gtRanks <- doGT(pheno.v = pheno,
                        data.m = bVals[, samps],
                        array = "EPIC",
                        ncores = 20)
        ebgs <- data.frame(doGSEAwt(rankEID.m = gtRanks, ptw.ls = entrezid,
                                    ncores = 20, minN = minsize, adjPVth = 1)$`Rank(P)`)
        toc(log = TRUE, quiet = TRUE)
        tmp <- strsplit2(tic.log(format = TRUE)[[1]], " ")
        log <- data.frame(method = gsub(":", "", tmp[1]), time = tmp[2], 
                          contrast = colnames(tfit$contrasts)[i])
        tic.clearlog()
        multi <- bind_rows(multi, log)
        
    }
    saveRDS(multi, file = inFile)
    
} else {
    multi <- readRDS(inFile)
    
}
```

Plot run-time results.

```{r}
multi %>% mutate(time = as.integer(time)) %>%
    mutate(cores = limma::strsplit2(method, ".", fixed=TRUE)[,2]) -> ebgseaMulti

p2 <- ggplot(ebgseaMulti, aes(x = cores, y = time/60, fill = contrast)) + 
    geom_bar(position = "dodge", stat = "identity") +
    labs(x = "No. cores", y = "Run time (minutes)", fill = "Contrast") +
    ggtitle("Using multiple cores for ebGSEA") 
p2
```

```{r}
inFile <- here("data/cache-runtime/run-time-mcores.rds")

if(!file.exists(inFile)){
    
    data("PathwayList")
    keep <- sapply(PathwayList, function(x) any(x %in% transIDs$symbol))
    symbol <- suppressMessages(lapply(PathwayList[keep], function(x){
        tmp <- unlist(transIDs$symbol[transIDs$symbol %in% x], use.names = FALSE)
        tmp[!is.na(tmp)]
    }))
    
    load(here("data/cache-intermediates/input.RData"))
    bVals <- ilogit2(mVals)
    anno <- minfi::getAnnotation(IlluminaHumanMethylationEPICanno.ilm10b4.hg19)
    multi <- NULL
    
    for(i in 1:ncol(tfit$contrasts)){
        
        tic("mgsa.glm.10")
        res <- methylglm(cpg.pval = tfit$p.value[,i],
                         FullAnnot = anno, minsize = minsize, maxsize = maxsize,
                         GS.list = symbol, GS.idtype = "SYMBOL", parallel = TRUE,
                         BPPARAM = BiocParallel::MulticoreParam(workers = 10))
        toc(log = TRUE, quiet = TRUE)
        tmp <- strsplit2(tic.log(format = TRUE)[[1]], " ")
        log <- data.frame(method = gsub(":", "", tmp[1]), time = tmp[2], 
                          contrast = colnames(tfit$contrasts)[i])
        tic.clearlog()
        multi <- bind_rows(multi, log)
        
        tic("mgsa.glm.20")
        res <- methylglm(cpg.pval = tfit$p.value[,i],
                         FullAnnot = anno, minsize = minsize, maxsize = maxsize,
                         GS.list = symbol, GS.idtype = "SYMBOL", parallel = TRUE,
                         BPPARAM = BiocParallel::MulticoreParam(workers = 20))
        toc(log = TRUE, quiet = TRUE)
        tmp <- strsplit2(tic.log(format = TRUE)[[1]], " ")
        log <- data.frame(method = gsub(":", "", tmp[1]), time = tmp[2], 
                          contrast = colnames(tfit$contrasts)[i])
        tic.clearlog()
        multi <- bind_rows(multi, log)
        
    }
    saveRDS(multi, file = inFile)
    
} else {
    multi <- readRDS(inFile)
}
```

Plot run-time results.

```{r}
multi %>% mutate(time = as.integer(time)) %>%
    mutate(cores = limma::strsplit2(method, ".", fixed=TRUE)[,3]) -> mglmMulti

p3 <- ggplot(mglmMulti, aes(x = cores, y = time/60, fill = contrast)) + 
    geom_bar(position = "dodge", stat = "identity") +
    labs(x = "No. cores", y = "Run time (minutes)", fill = "Contrast") +
    ggtitle("Using multiple cores for mGLM") 
p3
```

Combining single core and multi-core plots.

```{r}
p1 / (p2 | p3) + plot_layout(guides = "collect") &
  theme(text = element_text(size = 8))
```

# Run-time summary table

```{r}
timing$cores <- 1
timing$time <- as.integer(timing$time)
ebgseaMulti$cores <- as.integer(ebgseaMulti$cores)
mglmMulti$cores <- as.integer(mglmMulti$cores)
timing <- bind_rows(timing, ebgseaMulti,  mglmMulti)

timing$method[grepl("mgsa.glm",timing$method)] <- "mgsa.glm"
timing$method[grepl("ebgsea",timing$method)] <- "ebgsea"

timing %>% mutate(method = unname(dict[method])) %>% 
    group_by(method, cores) %>%
    mutate(time = as.integer(time)) %>%
    summarise(mean = mean(time)) %>%
    mutate(mins = mean/60) %>%
    dplyr::select(method, cores, mins) %>%
    arrange(cores, mins) %>%
    as_tibble() -> tab

tab %>% gt() %>%
    fmt_number(columns = vars(mins), decimals = 2) %>%
    cols_label(
        mins = md("**Minutes**"),
        cores = md("**Cores**"),
        method = md("**Method**")
        ) %>%
  tab_header(
    title = md("**Average run-time across all contrasts**"),
    subtitle = md("Using Broad MSigDB gene sets from `ChAMP` package")
    ) %>%
  tab_source_note(md("_*Blade Server: 24 CPUs - Intel(R) Xeon(R) Gold 6126 CPU @ 2.60GHz._"))
```
