==> analysis/08_methylGSAParamSweep.Rmd <==
---
title: 'Gene set testing for Illumina HumanMethylation Arrays'
subtitle: "Exploring minimum and maximum gene set size parameters for methylGSA"
author: "Jovana Maksimovic, Alicia Oshlack and Belinda Phipson"
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, message=FALSE, warning=FALSE, results='hide'}
library(here)
library(reshape2)
library(ggplot2)
library(glue)
library(tidyverse)
library(patchwork)
library(gt)
source(here("code/utility.R"))
```

# Explore effect of minimum and maximum gene set size parameters

We have run all three methylGSA methods, testing GO categories, on the three blood cell contrasts for the following combinations of minimum and maximum gene set size parameters: 

```{r}
params <- data.frame(minsize = c(1:5, rep(5, 5)),                      
                     maxsize = c(rep(5000, 5), seq(7000, 18000, by = 2750)))
params
```

As some of the methods take a considerable amount of time to perform the gene set testing analysis, we have created several scripts in order to run the analyses in parallel on a HPC.
The code used to run all the gene set testing analyses using the different methods can be found in the `code/param-sweep` directory. It consists of three scripts: `genSweepMethylGSA.R`, `paramSweepMethylGSA.R`, and `processSweepMethylGSA.R`. The `genSweepMethylGSA.R` script creates and submits [Slurm](https://slurm.schedmd.com/overview.html) job scripts that run the `paramSweepMethylGSA.R` script, for *each* combinations of parameters, in parallel, on a HPC. The results of each job are saved as an RDS file named `{method}.{minsz}.{maxsz}.rds` in the `output/methylgsa-params` directory. Once all jobs are complete, the `processSweepMethylGSA.R` must be executed to collate the results into a single object, which is then saved as `methylGSA-param-sweep.rds` in the `output/methylgsa-params` directory. The intermediate RDS files are moved into `output/methylgsa-params/.bin`, which can then be deleted, if no longer required. The subsequent section requires `methylGSA-param-sweep.rds` to be present in the `output/methylgsa-params` directory for downstream analysis and plotting.

## Load data

Read in the results of all the analyses.

```{r}
inFile <- here("output/methylgsa-params/methylGSA-param-sweep.rds")
dat <- readRDS(inFile)
```

## Examine results

```{r, message=FALSE, warning=FALSE, fig.height=7, fig.width=9}
dat %>% mutate(combo = glue("Min = {minsize}; Max = {maxsize}")) %>%
    mutate(method = unname((dict[method]))) %>%
    group_by(method, contrast, combo) %>%
    mutate(rank = 1:n()) %>%
    filter(rank <= 9) %>% 
    group_by(method, contrast) %>%
    mutate(params = factor(combo),
           params = factor(params,
                           levels = levels(params)[order(c(1:4,8:10,5:7))])) -> sub

methods <- sort(unique(sub$method))
p <- vector("list", length(methods) * length(unique(sub$contrast)))
i = 1
leg <- c(TRUE, rep(FALSE, length(p) - 1))

for(meth in methods) {
    for(j in 1:length(unique(sub$contrast))){
        sub %>% filter(method == meth & 
                           contrast == sort(unique(sub$contrast))[j]) -> subMeth
        
        p[[i]] <- ggplot(subMeth, aes(x=rank, y=Size, colour = params)) +
            geom_jitter(show.legend = leg[i], width = 0.15, height = 0.5,
                        shape = 20, alpha = 0.7) +
            scale_x_continuous(labels = 1:9, breaks = 1:9) +
            labs(x = "Rank", y = "No. genes in set",
                 colour = "Parameters") +
            guides(colour = guide_legend(ncol = 4, byrow = TRUE)) + 
            ggtitle(glue("{meth}: {sort(unique(sub$contrast))[j]}")) +
            theme(title = element_text(size = 8),
                  legend.text = element_text(size = 7),
                  legend.title = element_text(size = 8))
        i = i + 1
    }
}

p <- ((p[[1]] | p[[2]] | p[[3]]) /
    (p[[4]] | p[[5]] | p[[6]]) /
    (p[[7]] | p[[8]] | p[[9]])) + 
    plot_layout(guides = "collect") & 
    theme(legend.position = "bottom")
p
```

Save figure for use in manuscript.

```{r}
outDir <- here::here("output/figures")
if (!dir.exists(outDir)) dir.create(outDir)

fig <- here("output/figures/SFig-3E.rds")
saveRDS(p, fig, compress = FALSE)
```
